+++ server/libs/platform/platform-configuration/platform-configuration-api/src/main/java/com/bytechef/platform/configuration/notification/NotificationHandlerContext.java
@@ -16,7 +16,7 @@
 
 package com.bytechef.platform.configuration.notification;
 
-import com.bytechef.platform.configuration.domain.notification.Event;
+import com.bytechef.platform.configuration.domain.Event;
 
 /**
  * @author Matija Petanjek
