+++ server/libs/platform/platform-configuration/platform-configuration-rest/platform-configuration-rest-impl/src/main/java/com/bytechef/platform/configuration/web/rest/NotificationEventApiController.java
@@ -44,7 +44,7 @@ public class NotificationEventApiController implements NotificationEventApi {
     @Override
     public ResponseEntity<List<NotificationEventModel>> getNotificationEvents() {
         return ResponseEntity.ok(
-            eventService.findAll()
+            eventService.getEvents()
                 .stream()
                 .map(event -> conversionService.convert(event, NotificationEventModel.class))
                 .toList());
