+++ server/libs/platform/platform-configuration/platform-configuration-service/src/main/java/com/bytechef/platform/configuration/notification/WebhookNotificationSender.java
@@ -16,7 +16,7 @@
 
 package com.bytechef.platform.configuration.notification;
 
-import com.bytechef.platform.configuration.domain.notification.Notification;
+import com.bytechef.platform.configuration.domain.Notification;
 import org.springframework.stereotype.Component;
 
 /**
