+++ server/libs/platform/platform-configuration/platform-configuration-service/src/main/java/com/bytechef/platform/configuration/repository/NotificationRepository.java
@@ -16,7 +16,7 @@
 
 package com.bytechef.platform.configuration.repository;
 
-import com.bytechef.platform.configuration.domain.notification.Notification;
+import com.bytechef.platform.configuration.domain.Notification;
 import java.util.List;
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.ListCrudRepository;
